Converted insert action to scan
Trying to process every image in bulk while it's being added slows
things down too much. It also misses if an image is inserted as a
child element of the event (think AngularJS). By delaying it a bit
and checking after allows us to hopefully be a little less resource
intensive.